+++ source/reMatrix3.cpp
@@ -1,21 +1,11 @@
 //////////////////////////////////////////////////////////////////////////////////////////////////
 //
-// Name:        Razor Edge Math Library
-// File:        reMath.h
+// File:        reMatrix4.cpp
 // Project:     Razor Edge Classes
-// Description: Cummulative include of Razor Edge math classes
+// Description: Implementation of Matrix3 class (Math Module)
 // Copyright:   Copyright © 2004++ REGLabs
 // Author:      Pavel Chikul
 //
 //////////////////////////////////////////////////////////////////////////////////////////////////
 
-#ifndef __RE_MATH__
-#define __RE_MATH__
-
-#include "reVec2d.h"
-#include "reVec3d.h"
-#include "reMatrix4.h"
-#include "reQuaternion.h"
-#include "reMathUtil.h"
-
-#endif // __RE_MATH__
+#include "reMatrix3.h"