+++ CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+* Created an MSVS solution.
+* Unit test project and `Vec2d` coverage.
+
+### Changed
+* Renamed `Matrix` to `Matrix4` as we now have `Matrix3` in addition.
+* Updated `Vec2d` to the latest C++ standards.
+
 ## [1.0.1] - 22.07.2017
 ### Fixed
 * Inorrect definition of friend operators.